requirements/requirements-prod.txt
Django==3.0.5
django-environ==0.4.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
Pillow==7.1.1
django-sekizai==1.1.0
django-admin-honeypot==1.1.0
gunicorn==20.0.4
whitenoise==5.0.1
requests==2.23.0
django-heroku==0.3.1
sentry-sdk==0.14.3
django-health-check==3.12.1
loguru==0.4.1
